Add "Restart PowerShell Session" command
This change adds a new command to the extension, "Restart PowerShell
Session."  This command makes it possible for the user to restart the
current PowerShell session without restarting VS Code.  This is useful if
the user wants a fresh session or wants to switch from x64 to x86 or some
other build of PowerShell.

This change required a pretty large refactoring of the existing session
management code and the existing "features" so that they can be given the
new LanguageClient that gets created.